+++ R/utils.R
@@ -108,4 +108,3 @@ warnif <- function(..., msg) {
   }
 }
 
-